requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
requests==2.21.0
DerPyBooru==0.7.3
discord.py==0.16.12
discord.py[voice]==0.16.12
gunicorn==19.9.0
idna==2.8
multidict==4.5.2
pycparser==2.19
PyNaCl==1.3.0
six==1.12.0
urllib3==1.24.1
websockets==3.4
youtube-dl==2019.1.17